pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.123.10",
    "langchain>=1.1.2",
    "langchain-google-genai>=3.2.0",
    "langchain-ollama>=1.0.0",
    "langchain-postgres>=0.0.16",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "langgraph-checkpoint-postgres>=3.0.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "psycopg2>=2.9.11",
    "psycopg[binary,pool]>=3.3.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
    "uvicorn>=0.38.0",
]
